requirements.txt
click==8.0.1
colorama==0.4.4
Flask==2.0.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.2
python-dotenv==0.19.0
PyVISA==1.11.3
six==1.16.0
SQLAlchemy==1.4.25
typing-extensions==3.10.0.2
waitress==2.0.0
Werkzeug==2.0.2
psycopg2-binary==2.9.1
